requirements.txt
pip==20.0.2
Django==3.0.7
djangorestframework==3.11.0
pandas==1.0.1
numpy==1.18.1
gunicorn==20.0.4
dj-database-url==0.5.0
dj-static==0.0.6
docopt==0.6.2
gitignore==0.0.8
psycopg2==2.8.5
pytz==2019.3
static3==0.7.0
whitenoise==5.1.0
evds==0.2.0